backend/requirements.txt
google-auth
google-api-python-client
fastapi==0.115.5
uvicorn[standard]==0.30.1
google-cloud-aiplatform==1.70.0
google-cloud-storage==2.18.2
vertexai==1.70.0
python-dotenv==1.0.1


